ml_testing/requirements-train.txt
tensorflow
numpy
pandas
matplotlib
pillow
opencv-python
